Showing statistics on booking overview instead
