+++ carRental/src/components/Statistics/BookingStatistics.css
